[generator] Renamed from popularity to hierarchy.
