+++ generator/generator_tool/generator_tool.cpp
@@ -13,13 +13,13 @@
 #include "generator/feature_sorter.hpp"
 #include "generator/generate_info.hpp"
 #include "generator/geo_objects/geo_objects.hpp"
+#include "generator/hierarchy.hpp"
 #include "generator/locality_sorter.hpp"
 #include "generator/maxspeeds_builder.hpp"
 #include "generator/metalines_builder.hpp"
 #include "generator/osm_source.hpp"
 #include "generator/platform_helpers.hpp"
 #include "generator/popular_places_section_builder.hpp"
-#include "generator/popularity.hpp"
 #include "generator/regions/collector_region_info.hpp"
 #include "generator/regions/regions.hpp"
 #include "generator/restriction_generator.hpp"
