+++ generator/hierarchy.hpp
